+++ packages/database/CHANGELOG.md
@@ -1,8 +1,14 @@
 # CHANGELOG
 
-## 0.1.1 - 20200-08-18
+## 0.1.2 - 2020-08-18
 
-- Added `@fwl/tracing` to dependencies which was forgotten in 0.1.0
+- Added `DatabaseConfig` type and `defaultConfig` export
+- `connect` and `connectWithoutTracing` now takes a config as an optional argument and uses the `defaultConfig` if none is passed
+- Bumped `@fwl/tracing` to `0.3.1`
+
+## 0.1.1 - 2020-08-18
+
+- Added `@fwl/tracing` to dependencies which was forgotten in `0.1.0`
 
 ## 0.1.0 - 2020-08-12